pyproject.toml
[project]
name = "breast-cancer-wisconsin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "catboost>=1.2.8",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "streamlit>=1.45.1",
    "torch>=2.7.1",
    "xgboost>=3.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.13",
]


